Don't emit temps if we have reverted to our schedule
When a temp ended before the current scheduled basal was complete, it would revert back to the old basal rate for the schedule.  The old logic would consider this a new temp basal.  The logic currently checks to see if it matches the currently operating schedule and if it does, it ignores the event.