Handle constant basal schedules
If someone has a constant basal schedule, diasend will only give us one entry when that rate first goes into effect.  We do not know anything about when it might've changed or anything like that.  But, we know that changing an injection set causes the pump to go to a rate of 0, assuming that injection sets are changed every 5 days, we limit the length of basals to 5 days.

Also, it is possible to have a schedule that is just the same rate over and over and over again, but set to start and end at different times.  We should consider this as just one rate and merge them together.  Because otherwise, we do not generate all scheduled basals.