Do not fabricate scheduled basals
Also, do not update the schedule based on basals that we have seen.

This change implements both of these things, but there is still a case where we need to fabricate scheduled basals and that is when a temp percent is active.  This change does not implement that yet, but it has adjusted the tests (specifically withoutSettingsChanges) to expect that.  So what's left is to make that test pass.