Update tests, moment-timezone and minor fixes
This commit includes updates to the staticBasal and withoutSettingsChanges tests.  withSettingsChanges will also need to change, but it is largely equivalent to withoutSettingsChanges, so I have manually adjusted the expected output there and then will do a strict check of what is generated against the current expected output of withSettingsChanges in order to ensure that things are lining up.

Committing to get this stuff in git.